+++ src/main/java/com/isep/game/cards/ProgressToken.java
@@ -1,4 +1,4 @@
-package com.isep;
+package com.isep.game.cards;
 
 /**
  * A class representing a progress token.
