+++ src/main/java/com/isep/game/wonders/Ephesus.java
@@ -1,31 +1,33 @@
-package com.isep;
+package com.isep.game.wonders;
+
+import com.isep.game.cards.*;
 
 import java.util.ArrayList;
 
-public class Rhodes extends Wonder
+public class Ephesus extends Wonder
 {
     // Constructor
-    public Rhodes()
+    public Ephesus()
     {
-        super("Rhodes", new ArrayList<Stage>() {{
-            add(new Stage(4, 2, false, false, false, 1));
-            add(new Stage(4, 2, true, true, false, 1));
-            add(new Stage(5, 3, false, false, false, 2));
-            add(new Stage(6, 3, true, true, false, 3));
-            add(new Stage(7, 4, false, false, false, 4));
+        super("Ephesus", new ArrayList<Stage>() {{
+            add(new Stage(3, 2, false, false, false, 1));
+            add(new Stage(3, 2, true, true, false, 2));
+            add(new Stage(4, 3, false, true, false, 2));
+            add(new Stage(5, 3, true, true, false, 2));
+            add(new Stage(7, 4, false, false, false, 3));
         }}, new Deck());
         this.buildDeck();
         this.deck.shuffle();
     }
 
     /**
-     * Creates the {@link Deck} associated to the {@link Rhodes} wonder.
+     * Creates the {@link Deck} associated to the {@link Ephesus} wonder.
      * @author Minaé RAFFIN
      */
     private void buildDeck()
     {
         // Yellow cards
-        for(int i = 1 ; i <= 2; i++)
+        for(int i = 1 ; i <= 3; i++)
             this.deck.addCard(new YellowCard());
         // Grey cards
         for(int i = 1 ; i <= 2; i++)
@@ -39,14 +41,14 @@ public class Rhodes extends Wonder
         for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new GreyCard(GreyCard.Material.PAPYRUS));
         // Green cards
-        for(int i = 1 ; i <= 2; i++)
+        for(int i = 1 ; i <= 1; i++)
             this.deck.addCard(new GreenCard(GreenCard.ScienceSymbol.GEAR));
-        for(int i = 1 ; i <= 1; i++)
+        for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new GreenCard(GreenCard.ScienceSymbol.COMPASS));
-        for(int i = 1 ; i <= 1; i++)
+        for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new GreenCard(GreenCard.ScienceSymbol.TABLET));
         // Blue cards
-        for(int i = 1 ; i <= 2; i++)
+        for(int i = 1 ; i <= 1; i++)
             this.deck.addCard(new BlueCard(3, false));
         for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new BlueCard(2, true));
@@ -55,7 +57,7 @@ public class Rhodes extends Wonder
             this.deck.addCard(new RedCard(0));
         for(int i = 1 ; i <= 1; i++)
             this.deck.addCard(new RedCard(1));
-        for(int i = 1 ; i <= 2; i++)
+        for(int i = 1 ; i <= 1; i++)
             this.deck.addCard(new RedCard(2));
     }
 }
