+++ src/main/java/com/isep/game/wonders/Halicarnassus.java
@@ -1,25 +1,27 @@
-package com.isep;
+package com.isep.game.wonders;
+
+import com.isep.game.cards.*;
 
 import java.util.ArrayList;
 
-public class Giza extends Wonder
+public class Halicarnassus extends Wonder
 {
     // Constructor
-    public Giza()
+    public Halicarnassus()
     {
-        super("Giza", new ArrayList<Stage>() {{
-            add(new Stage(4, 2, false, false, false, 1));
-            add(new Stage(5, 2, true, false, false, 2));
+        super("Halicarnassus", new ArrayList<Stage>() {{
+            add(new Stage(3, 2, false, false, false, 1));
+            add(new Stage(3, 2, true, true, false, 2));
             add(new Stage(6, 3, false, false, false, 3));
-            add(new Stage(7, 3, true, false, false, 4));
-            add(new Stage(8, 4, false, false, false, 5));
+            add(new Stage(5, 3, true, true, false, 3));
+            add(new Stage(7, 4, false, false, false, 4));
         }}, new Deck());
         this.buildDeck();
         this.deck.shuffle();
     }
 
     /**
-     * Creates the {@link Deck} associated to the {@link Giza} wonder.
+     * Creates the {@link Deck} associated to the {@link Halicarnassus} wonder.
      * @author Minaé RAFFIN
      */
     private void buildDeck()
@@ -30,30 +32,30 @@ public class Giza extends Wonder
         // Grey cards
         for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new GreyCard(GreyCard.Material.STONE));
-        for(int i = 1 ; i <= 1; i++)
+        for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new GreyCard(GreyCard.Material.BRICK));
         for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new GreyCard(GreyCard.Material.WOOD));
         for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new GreyCard(GreyCard.Material.GLASS));
-        for(int i = 1 ; i <= 2; i++)
+        for(int i = 1 ; i <= 1; i++)
             this.deck.addCard(new GreyCard(GreyCard.Material.PAPYRUS));
         // Green cards
         for(int i = 1 ; i <= 1; i++)
             this.deck.addCard(new GreenCard(GreenCard.ScienceSymbol.GEAR));
-        for(int i = 1 ; i <= 2; i++)
-            this.deck.addCard(new GreenCard(GreenCard.ScienceSymbol.COMPASS));
         for(int i = 1 ; i <= 1; i++)
+            this.deck.addCard(new GreenCard(GreenCard.ScienceSymbol.COMPASS));
+        for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new GreenCard(GreenCard.ScienceSymbol.TABLET));
         // Blue cards
         for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new BlueCard(3, false));
-        for(int i = 1 ; i <= 3; i++)
+        for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new BlueCard(2, true));
         // Red cards
         for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new RedCard(0));
-        for(int i = 1 ; i <= 0; i++)
+        for(int i = 1 ; i <= 1; i++)
             this.deck.addCard(new RedCard(1));
         for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new RedCard(2));
