+++ src/main/java/com/isep/game/wonders/Olympia.java
@@ -1,16 +1,18 @@
-package com.isep;
+package com.isep.game.wonders;
+
+import com.isep.game.cards.*;
 
 import java.util.ArrayList;
 
-public class Halicarnassus extends Wonder
+public class Olympia extends Wonder
 {
     // Constructor
-    public Halicarnassus()
+    public Olympia()
     {
-        super("Halicarnassus", new ArrayList<Stage>() {{
+        super("Olympia", new ArrayList<Stage>() {{
             add(new Stage(3, 2, false, false, false, 1));
-            add(new Stage(3, 2, true, true, false, 2));
-            add(new Stage(6, 3, false, false, false, 3));
+            add(new Stage(2, 2, true, true, false, 2));
+            add(new Stage(5, 3, false, false, false, 2));
             add(new Stage(5, 3, true, true, false, 3));
             add(new Stage(7, 4, false, false, false, 4));
         }}, new Deck());
@@ -19,7 +21,7 @@ public class Halicarnassus extends Wonder
     }
 
     /**
-     * Creates the {@link Deck} associated to the {@link Halicarnassus} wonder.
+     * Creates the {@link Deck} associated to the {@link Olympia} wonder.
      * @author Minaé RAFFIN
      */
     private void buildDeck()
@@ -32,11 +34,11 @@ public class Halicarnassus extends Wonder
             this.deck.addCard(new GreyCard(GreyCard.Material.STONE));
         for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new GreyCard(GreyCard.Material.BRICK));
-        for(int i = 1 ; i <= 2; i++)
+        for(int i = 1 ; i <= 1; i++)
             this.deck.addCard(new GreyCard(GreyCard.Material.WOOD));
         for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new GreyCard(GreyCard.Material.GLASS));
-        for(int i = 1 ; i <= 1; i++)
+        for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new GreyCard(GreyCard.Material.PAPYRUS));
         // Green cards
         for(int i = 1 ; i <= 1; i++)
@@ -46,9 +48,9 @@ public class Halicarnassus extends Wonder
         for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new GreenCard(GreenCard.ScienceSymbol.TABLET));
         // Blue cards
-        for(int i = 1 ; i <= 2; i++)
+        for(int i = 1 ; i <= 1; i++)
             this.deck.addCard(new BlueCard(3, false));
-        for(int i = 1 ; i <= 2; i++)
+        for(int i = 1 ; i <= 3; i++)
             this.deck.addCard(new BlueCard(2, true));
         // Red cards
         for(int i = 1 ; i <= 2; i++)
@@ -59,3 +61,4 @@ public class Halicarnassus extends Wonder
             this.deck.addCard(new RedCard(2));
     }
 }
+
