+++ src/main/java/com/isep/game/wonders/Rhodes.java
@@ -1,17 +1,19 @@
-package com.isep;
+package com.isep.game.wonders;
+
+import com.isep.game.cards.*;
 
 import java.util.ArrayList;
 
-public class Babylon extends Wonder
+public class Rhodes extends Wonder
 {
     // Constructor
-    public Babylon()
+    public Rhodes()
     {
-        super("Babylon", new ArrayList<Stage>() {{
-            add(new Stage(3, 2, false, false, false, 1));
-            add(new Stage(0, 2, true, true, false, 2));
-            add(new Stage(5, 3, false, false, false, 3));
-            add(new Stage(5, 3, true, true, false, 4));
+        super("Rhodes", new ArrayList<Stage>() {{
+            add(new Stage(4, 2, false, false, false, 1));
+            add(new Stage(4, 2, true, true, false, 1));
+            add(new Stage(5, 3, false, false, false, 2));
+            add(new Stage(6, 3, true, true, false, 3));
             add(new Stage(7, 4, false, false, false, 4));
         }}, new Deck());
         this.buildDeck();
@@ -19,16 +21,16 @@ public class Babylon extends Wonder
     }
 
     /**
-     * Creates the {@link Deck} associated to the {@link Babylon} wonder.
+     * Creates the {@link Deck} associated to the {@link Rhodes} wonder.
      * @author Minaé RAFFIN
      */
     private void buildDeck()
     {
         // Yellow cards
-        for(int i = 1 ; i <= 3; i++)
+        for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new YellowCard());
         // Grey cards
-        for(int i = 1 ; i <= 1; i++)
+        for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new GreyCard(GreyCard.Material.STONE));
         for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new GreyCard(GreyCard.Material.BRICK));
@@ -41,7 +43,7 @@ public class Babylon extends Wonder
         // Green cards
         for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new GreenCard(GreenCard.ScienceSymbol.GEAR));
-        for(int i = 1 ; i <= 2; i++)
+        for(int i = 1 ; i <= 1; i++)
             this.deck.addCard(new GreenCard(GreenCard.ScienceSymbol.COMPASS));
         for(int i = 1 ; i <= 1; i++)
             this.deck.addCard(new GreenCard(GreenCard.ScienceSymbol.TABLET));
@@ -55,7 +57,7 @@ public class Babylon extends Wonder
             this.deck.addCard(new RedCard(0));
         for(int i = 1 ; i <= 1; i++)
             this.deck.addCard(new RedCard(1));
-        for(int i = 1 ; i <= 1; i++)
+        for(int i = 1 ; i <= 2; i++)
             this.deck.addCard(new RedCard(2));
     }
 }
